Refactogin + Authentication

+++ src/repository/user.js
@@ -1,7 +1,7 @@
 const {getKnex, tables} = require("../data");
 const getChildLogger = require("../core/logging").getChildLogger;
-
-
+const {verifyPassword} = require("../core/password");
+const ServiceError = require("../core/serviceError");
 
 
 const login = (username, password) => {
@@ -27,7 +27,8 @@ const create = async ({email,username,firstname,lastname,password_hash,roles}) =
   }catch (error){
     const logger = getChildLogger("users-repo");
     logger.error("Error in create",{error});
-    throw error;
+    throw ServiceError.unknown("This should not have happened/ wrong parameters?");
+
   }
 
 }
@@ -60,45 +61,48 @@ const findById = (id) => {
   .first();
 }
 const updateById = async (id, {...user}) => {
-
-  const check = await getKnex()(tables.users)
-    .select()
-    .where("id", id)
-    .first();
-  
+  const check = await findById(id);
   try{
     if(check){
-      await getKnex()(tables.users)
-      .update({
-        email:user.email,
-        username:user.username,
-        firstname:user.firstname,
-        lastname:user.lastname,
-        password_hash:user.password_hash,
-      })
-      .where("id",id);
-      return await findById(id);
-    }else{
-       return 0;
-    }
-    
+        const passwordCorrect = await verifyPassword(user.password, check.password_hash);
+        if(passwordCorrect){
+          await getKnex()(tables.users)
+          .update({
+            email:user.email,
+            username:user.username,
+            firstname:user.firstname,
+            lastname:user.lastname,
+            password_hash:user.password_hash,
+          })
+          .where("id",id);
+          return await findById(id);
+        }
+        throw ServiceError.unauthorized(`Not possible to change someone else his/her profile.`);    
+      }
+    else{
+      throw ServiceError.notFound(`No user found with id: ${id} `);
+    }    
  } catch (error){
     const logger = getChildLogger("users-repo");
     logger.error("Error in updateBy", error);
-    throw error;
+    throw ServiceError.unknown("This should not have happened/ wrong parameters?");
   }
 }
 
 const deleteById = async (id) => {
-  try {
+  const check = await findById(id);
+  try{
+    if(check){
     const rowsAffected = await getKnex()(tables.users)
     .delete()
     .where("id",id);
     return rowsAffected > 0;
+    }
   }catch (error) {
     const logger = getChildLogger("users-repo");
     logger.error("Error in deleteby", {error});
-    throw error;
+    throw ServiceError.unknown("This should not have happened/ wrong parameters?");
+
   }
 }
